Addressing PR comments, moving some common template parsing methods to a common TemplateUtil class
