+++ src/main/java/org/opensearch/flowframework/model/Template.java
@@ -25,6 +25,8 @@ import java.util.Map;
 import java.util.Map.Entry;
 
 import static org.opensearch.core.xcontent.XContentParserUtils.ensureExpectedToken;
+import static org.opensearch.flowframework.common.TemplateUtil.jsonToParser;
+import static org.opensearch.flowframework.common.TemplateUtil.parseStringToStringMap;
 
 /**
  * The Template is the central data structure which configures workflows. This object is used to parse JSON communicated via REST API.
@@ -523,23 +525,6 @@ public class Template implements ToXContentObject {
     }
 
     /**
-     * Converts a JSON string into an XContentParser
-     *
-     * @param json the json string
-     * @return The XContent parser for the json string
-     * @throws IOException on failure to create the parser
-    */
-    public static XContentParser jsonToParser(String json) throws IOException {
-        XContentParser parser = JsonXContent.jsonXContent.createParser(
-            NamedXContentRegistry.EMPTY,
-            LoggingDeprecationHandler.INSTANCE,
-            json
-        );
-        ensureExpectedToken(XContentParser.Token.START_OBJECT, parser.nextToken(), parser);
-        return parser;
-    }
-
-    /**
      * Parse a JSON use case template
      *
      * @param json A string containing a JSON representation of a use case template
@@ -554,39 +539,6 @@ public class Template implements ToXContentObject {
         );
         ensureExpectedToken(XContentParser.Token.START_OBJECT, parser.nextToken(), parser);
         return parse(parser);
-    }
-
-    /**
-     * Builds an XContent object representing a map of String keys to String values.
-     *
-     * @param xContentBuilder An XContent builder whose position is at the start of the map object to build
-     * @param map A map as key-value String pairs.
-     * @throws IOException on a build failure
-     */
-    public static void buildStringToStringMap(XContentBuilder xContentBuilder, Map<?, ?> map) throws IOException {
-        xContentBuilder.startObject();
-        for (Entry<?, ?> e : map.entrySet()) {
-            xContentBuilder.field((String) e.getKey(), (String) e.getValue());
-        }
-        xContentBuilder.endObject();
-    }
-
-    /**
-     * Parses an XContent object representing a map of String keys to String values.
-     *
-     * @param parser An XContent parser whose position is at the start of the map object to parse
-     * @return A map as identified by the key-value pairs in the XContent
-     * @throws IOException on a parse failure
-     */
-    public static Map<String, String> parseStringToStringMap(XContentParser parser) throws IOException {
-        ensureExpectedToken(XContentParser.Token.START_OBJECT, parser.currentToken(), parser);
-        Map<String, String> map = new HashMap<>();
-        while (parser.nextToken() != XContentParser.Token.END_OBJECT) {
-            String fieldName = parser.currentName();
-            parser.nextToken();
-            map.put(fieldName, parser.text());
-        }
-        return map;
     }
 
     /**
